% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deflection}
\alias{deflection}
\title{Road deflection dataset.}
\format{\code{deflection}: A data frame with 1000 rows and 2 variables.

\itemize{
  \item SLK. Spatial location of data. SLK is short for the straight line kilometer.
  \item Deflection. The monitored road deflection value.
}}
\usage{
deflection
}
\description{
The "deflection" dataset is a sample of the road deflection data
             monitored and collected by Main Roads Western Australia.
}
\author{
Yongze Song \email{yongze.song@postgrad.curtin.edu.au}
}
\keyword{dataset}
\keyword{deflection}
