% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segid.matrix}
\alias{segid.matrix}
\title{A matrix of segmentations with different methods for data "tsdwa".}
\format{\code{segid.matrix}: A matrix with 300 rows and 3 columes,
        representing segmentations of three methods, CDA, MCV and SHS.}
\usage{
segid.matrix
}
\description{
Segmentation results of CDA, MCV and SHS methods for data "tsdwa".
}
\author{
Yongze Song \email{yongze.song@postgrad.curtin.edu.au}
}
\keyword{dataset}
\keyword{segmentations}
