% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsdwa}
\alias{tsdwa}
\title{Traffic speed deflectometer (TSD) data of pavement deteriorations.}
\format{\code{tsdwa}: A data frame with 5000 rows and 8 variables.

\itemize{
  \item id. Number of observation.
  \item SLK.start. Spatial start location of data. SLK is short for the straight line kilometer.
  \item SLK.end. Spatial end location of data.
  \item SurfType. Surfacing type.
  \item PvtType. Pavement type.
  \item Curvature.
  \item Deflection.
  \item BLI. Base layer index.
}}
\usage{
tsdwa
}
\description{
The "tsdwa" dataset is a sample of the pavement deterioration data
             monitored and collected by Main Roads Western Australia.
}
\author{
Yongze Song \email{yongze.song@postgrad.curtin.edu.au}
}
\keyword{TSD}
\keyword{dataset}
