% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Preprocessing for field monitoring data.}
\usage{
preprocessing(var = "deflection", location = "SLK", data = data)
}
\arguments{
\item{var}{A character of the name of a variable in a dataset,
such as a road pavement performance indicator.}

\item{location}{A character of the name of spatial locations in a dataset.}

\item{data}{A data frame of monitoring data.}
}
\description{
The preprocessing includes two steps:
             removing missing data and ordering data by spatial locations.
}
\examples{
testdata <- tsdwa[1:100,]
testdata <- preprocessing(var = "Deflection", location = "SLK.start", data = testdata)

}
